feat(HNG-106): moved the blogcommentsection from routes to component/common
